updates ckeytools
since the cog we were developing is supposed to be ckeytool's new version
