+++ ckeytools/helpers.py
