--- ss13_verify/info.json
@@ -1,26 +1,0 @@
-{
-    "author": [
-        "Mal0"
-    ],
-    "name": "SS13Verify",
-    "short": "SS13 Discord verification system",
-    "description": "A comprehensive SS13 Discord verification system with ticket-based verification, auto-verification, role assignment, and database linking. Replaces tgverify and tgdb with modern Discord features.",
-    "install_msg": "Thank you for installing SS13Verify! Use [p]ss13verify to configure the verification system. Make sure to set up your database connection and configure the verification panel.",
-    "requirements": [
-        "aiomysql",
-        "sqlalchemy[asyncio]"
-    ],
-    "tags": [
-        "SS13",
-        "Space Station 13",
-        "verification",
-        "database",
-        "tickets",
-        "roles",
-        "ckey"
-    ],
-    "min_bot_version": "3.5.0",
-    "hidden": false,
-    "end_user_data_statement": "This cog stores user verification data including Discord IDs, ckeys, verification tokens, and timestamps in both the configured MySQL database and Red's Config system. It also tracks open tickets and deverified users.",
-    "type": "COG"
-}
